+++ common/serverCard.cpp
@@ -7,12 +7,12 @@ ServerCard::ServerCard(std::shared_ptr<CardInfo> info, ServerCardZone *zone)
     mSoul = info->soul();
 }
 
-ServerCard::ServerCard(size_t pos, ServerCardZone *zone)
+ServerCard::ServerCard(int pos, ServerCardZone *zone)
     : mZone(zone) {
     setPos(pos);
 }
 
-void ServerCard::setPos(size_t pos) {
+void ServerCard::setPos(int pos) {
     mPosition = pos;
     if (pos < 3)
         mRow = CenterStage;
