+++ common/serverPlayer.h
@@ -23,7 +23,7 @@ class ServerPlayer
 {
     ServerGame *mGame;
     ServerProtocolHandler *mClient;
-    size_t mId;
+    int mId;
     bool mReady = false;
     bool mMulliganFinished = false;
     bool mStartingPlayer = false;
@@ -34,22 +34,22 @@ class ServerPlayer
     int mLevel = 0;
 
 public:
-    ServerPlayer(ServerGame *game, ServerProtocolHandler *client, size_t id);
+    ServerPlayer(ServerGame *game, ServerProtocolHandler *client, int id);
 
-    size_t id() const { return mId; }
+    int id() const { return mId; }
     bool ready() const { return mReady; }
     void setReady(bool ready) { mReady = ready; }
     bool mulliganFinished() const { return mMulliganFinished; }
     bool startingPlayer() const { return mStartingPlayer; }
     void setStartingPlayer() { mStartingPlayer = true; }
-    bool hasBattleOpponent(size_t pos) const;
+    bool hasBattleOpponent(int pos) const;
 
     void clearExpectedComands();
     void addExpectedCommand(const std::string &command);
     bool expectsCommand(const GameCommand &command);
 
     void processGameCommand(GameCommand &cmd);
-    void sendGameEvent(const ::google::protobuf::Message &event, size_t playerId = 0);
+    void sendGameEvent(const ::google::protobuf::Message &event, int playerId = 0);
 
     void addDeck(const std::string &deck);
     DeckList* deck() { return mDeck.get(); }
@@ -60,12 +60,12 @@ public:
     void startTurn();
     void dealStartingHand();
     void mulligan(const CommandMulligan &cmd);
-    void drawCards(size_t number);
-    void moveCard(std::string_view startZoneName, const std::vector<size_t> &cardIds, std::string_view targetZoneName);
+    void drawCards(int number);
+    void moveCard(std::string_view startZoneName, const std::vector<int> &cardIds, std::string_view targetZoneName);
     void processClockPhaseResult(const CommandClockPhase &cmd);
     void playCard(const CommandPlayCard &cmd);
     void playCharacter(const CommandPlayCard &cmd);
-    void playClimax(size_t handIndex);
+    void playClimax(int handIndex);
     void switchPositions(const CommandSwitchStagePositions &cmd);
     bool canPlay(ServerCard *card);
     void climaxPhase();
